+++ health-services/individual/src/main/java/org/egov/individual/web/models/IndividualSearch.java
@@ -39,7 +39,7 @@ public class IndividualSearch   {
     private List<String> id = null;
 
     @JsonProperty("individualId")
-    private String individualId = null;
+    private List<String> individualId = null;
 
     @JsonProperty("clientReferenceId")
     private List<String> clientReferenceId = null;
@@ -57,7 +57,7 @@ public class IndividualSearch   {
     private Gender gender = null;
 
     @JsonProperty("mobileNumber")
-    private String mobileNumber = null;
+    private List<String> mobileNumber = null;
 
     @JsonProperty("socialCategory")
     @Exclude
@@ -94,11 +94,11 @@ public class IndividualSearch   {
 
     @Exclude
     @JsonProperty("username")
-    private String username;
+    private List<String> username;
 
     @Exclude
     @JsonProperty("userId")
-    private Long userId;
+    private List<Long> userId;
 
     @Exclude
     @JsonProperty("userUuid")