+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/individual/IndividualSearch.java
@@ -33,7 +33,7 @@ public class IndividualSearch   {
     private List<String> id = null;
 
     @JsonProperty("individualId")
-    private String individualId = null;
+    private List<String> individualId = null;
 
     @JsonProperty("clientReferenceId")
     private List<String> clientReferenceId = null;
@@ -51,7 +51,7 @@ public class IndividualSearch   {
     private Gender gender = null;
 
     @JsonProperty("mobileNumber")
-    private String mobileNumber = null;
+    private List<String> mobileNumber = null;
 
     @JsonProperty("socialCategory")
     private String socialCategory = null;
@@ -79,10 +79,10 @@ public class IndividualSearch   {
     private List<String> roleCodes = null;
 
     @JsonProperty("username")
-    private String username;
+    private List<String> username;
 
     @JsonProperty("userId")
-    private Long userId;
+    private List<Long> userId;
 
     @JsonProperty("userUuid")
     @Size(min = 1)