+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/project/ProjectBeneficiarySearch.java
@@ -1,5 +1,8 @@
 package org.egov.common.models.project;
 
+import java.util.List;
+import javax.validation.constraints.Size;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
@@ -8,9 +11,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.validation.annotation.Validated;
-
-import javax.validation.constraints.Size;
-import java.util.List;
 
 /**
 * Search model for project beneficiary.
@@ -34,12 +34,10 @@ public class ProjectBeneficiarySearch {
     private String tenantId = null;
 
     @JsonProperty("projectId")
-    @Size(min=2,max=64)
-    private String projectId = null;
+    private List<String> projectId = null;
 
     @JsonProperty("beneficiaryId")
-    @Size(min=2,max=64)
-    private String beneficiaryId = null;
+    private List<String> beneficiaryId = null;
 
     @JsonProperty("clientReferenceId")
     private List<String> clientReferenceId = null;