+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/stock/StockReconciliationSearch.java
@@ -1,5 +1,8 @@
 package org.egov.common.models.stock;
 
+import java.util.List;
+import javax.validation.Valid;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -7,10 +10,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.validation.annotation.Validated;
-
-import javax.validation.Valid;
-import javax.validation.constraints.Size;
-import java.util.List;
 
 /**
  * StockReconciliationSearch
@@ -32,11 +31,9 @@ public class StockReconciliationSearch {
     private List<String> clientReferenceId = null;
 
     @JsonProperty("facilityId")
-    @Size(min=2, max=64)
-    private String facilityId = null;
+    private List<String> facilityId = null;
 
     @JsonProperty("productVariantId")
-    @Size(min=2, max=64)
-    private String productVariantId = null;
+    private List<String> productVariantId = null;
 }