+++ health-services/product/src/test/java/org/egov/product/web/controllers/ProductVariantApiControllerTest.java
@@ -26,6 +26,7 @@ import org.springframework.test.web.servlet.MvcResult;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -186,7 +187,7 @@ public class ProductVariantApiControllerTest {
     void shouldAcceptSearchRequestAndReturnProductsVariants() throws Exception {
 
         ProductVariantSearchRequest productVariantSearchRequest = ProductVariantSearchRequest.builder().productVariant(
-                ProductVariantSearch.builder().productId("101").build()
+                ProductVariantSearch.builder().productId(Arrays.asList("101","102")).build()
         ).requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build()).build();
         when(productVariantService.search(any(ProductVariantSearchRequest.class),
                 any(Integer.class),
@@ -213,7 +214,7 @@ public class ProductVariantApiControllerTest {
     void shouldThrowExceptionIfNoResultFound() throws Exception {
 
         ProductVariantSearchRequest productVariantSearchRequest = ProductVariantSearchRequest.builder().productVariant(
-                ProductVariantSearch.builder().productId("101").build()
+                ProductVariantSearch.builder().productId(Arrays.asList("101","102")).build()
         ).requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build()).build();
         when(productVariantService.search(any(ProductVariantSearchRequest.class),
                 any(Integer.class),