+++ health-services/referralmanagement/src/main/java/org/egov/referralmanagement/service/DownsyncService.java
@@ -261,7 +261,7 @@ public class DownsyncService {
 	        		
 	        ProjectBeneficiarySearch search = ProjectBeneficiarySearch.builder()
 					.id(ids)
-					.projectId(downsyncRequest.getDownsyncCriteria().getProjectId())
+					.projectId(Collections.singletonList(downsyncRequest.getDownsyncCriteria().getProjectId()))
 					.build();
 			
 			BeneficiarySearchRequest searchRequest = BeneficiarySearchRequest.builder()
@@ -305,7 +305,7 @@ public class DownsyncService {
 	        		
 	        TaskSearch search = TaskSearch.builder()
 					.id(taskIds)
-					.projectId(downsyncRequest.getDownsyncCriteria().getProjectId())
+					.projectId(Collections.singletonList(downsyncRequest.getDownsyncCriteria().getProjectId()))
 					.build();
 			
 			TaskSearchRequest searchRequest = TaskSearchRequest.builder()