+++ health-services/stock/src/main/java/org/egov/stock/web/models/StockSearch.java
@@ -40,14 +40,13 @@ public class StockSearch {
     private String facilityId = null;
 
     @JsonProperty("productVariantId")
-    @Size(min=2, max=64)
-    private String productVariantId = null;
+    private List<String> productVariantId = null;
 
     @JsonProperty("referenceId")
     private String referenceId = null;
 
     @JsonProperty("wayBillNumber")
-    private String wayBillNumber = null;
+    private List<String> wayBillNumber = null;
 
     @JsonProperty("referenceIdType")
     @Size(min=2, max=64)